fix error about qloadlen

+++ solve_distribute/solve.c
@@ -3,7 +3,7 @@
 #include <glpk.h>
 
 int main(int argc,char *argv[]){
-  int q[] = {atof(argv[1]),atof(argv[2]),atof(argv[3])};    //queue length
+  int q[] = {atoi(argv[1]),atoi(argv[2]),atoi(argv[3])};    //queue length
   int n[] = {atoi(argv[4]),atoi(argv[5]),atoi(argv[6])};
   int cmax = atof(argv[7]);    //amount
   glp_prob *mip = glp_create_prob();
